Add menu, update readme
