+++ CMakeLists.txt
@@ -1,10 +1,20 @@
 cmake_minimum_required(VERSION 3.16)
-project(pixhawk_heartbeat_detector CXX)
+project(orbis CXX)
 
 set(CMAKE_CXX_STANDARD 17)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 find_package(MAVSDK REQUIRED)
 
-add_executable(main main.cpp)
-target_link_libraries(main PRIVATE MAVSDK::mavsdk)
+add_executable(orbis
+    src/main.cpp
+    src/monitor.cpp
+)
+
+target_include_directories(orbis PRIVATE
+    ${CMAKE_CURRENT_SOURCE_DIR}/include
+)
+
+target_link_libraries(orbis PRIVATE
+    MAVSDK::mavsdk
+)
